test: move discourse stubs to helpers
